convert support-utilities to typescript
